+++ integration-tests/cli/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @openfn/integration-tests-cli
 
+## 1.0.6
+
+### Patch Changes
+
+- Update dependencies
+- Updated dependencies
+  - @openfn/lightning-mock@2.2.4
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [c47b65f]
+  - @openfn/lightning-mock@2.2.3
+
 ## 1.0.4
 
 ### Patch Changes