+++ packages/describe-package/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/describe-package
 
+## 0.1.5
+
+### Patch Changes
+
+- Update dependencies
+
 ## 0.1.4
 
 ### Patch Changes