+++ packages/lightning-mock/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @openfn/lightning-mock
 
+## 2.2.4
+
+### Patch Changes
+
+- Update dependencies
+- Updated dependencies
+  - @openfn/engine-multi@1.6.9
+  - @openfn/lexicon@1.2.3
+  - @openfn/runtime@1.7.2
+  - @openfn/logger@1.0.6
+
+## 2.2.3
+
+### Patch Changes
+
+- c47b65f: Ensure json payloads respect sockey delay
+
 ## 2.2.2
 
 ### Patch Changes