+++ packages/project/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/project
 
+## 0.1.1
+
+### Patch Changes
+
+- Update dependencies
+- Updated dependencies
+  - @openfn/lexicon@1.2.3
+  - @openfn/logger@1.0.6
+
 ## 0.1.0
 
 - Initial release of project, with basic capability to sync fs and app state